QnA/MediaSleuth360/requirements.txt
groq
moviepy==1.0.3
streamlit
google.generativeai
IPython

